evaluators package moved, crossvalidator merged into croassvalidationevaluator, evaluator changed from interface to abstract class
