+++ src/main/java/evaluation/storage/ClassifierResults.java
@@ -12,7 +12,7 @@
  *   You should have received a copy of the GNU General Public License
  *   along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package evaluation;
+package evaluation.storage;
 
 import fileIO.OutFile;
 import java.io.File;
