+++ src/main/java/evaluation/storage/ClassifierResultsCollection.java
@@ -1,4 +1,4 @@
-package evaluation;
+package evaluation.storage;
 
 import java.io.FileNotFoundException;
 import utilities.DebugPrinting;
