+++ src/main/java/evaluation/storage/MultipleClassifierResultsCollection.java
@@ -1,6 +1,6 @@
-package evaluation;
+package evaluation.storage;
 
-import evaluation.ClassifierResultsCollection;
+import evaluation.storage.ClassifierResultsCollection;
 import experiments.DataSets;
 import evaluation.MultipleClassifierEvaluation;
 import java.io.File;
