+++ src/main/java/timeseriesweka/classifiers/ElasticEnsemble.java
@@ -35,7 +35,7 @@ import weka.core.Instance;
 import weka.core.Instances;
 import timeseriesweka.filters.DerivativeFilter;
 import utilities.WritableTestResults;
-import evaluation.ClassifierResults;
+import evaluation.storage.ClassifierResults;
 import weka.core.TechnicalInformation;
 
 /**
