+++ src/main/java/timeseriesweka/classifiers/SlowDTW_1NN.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import timeseriesweka.elastic_distance_measures.DTW;
 import timeseriesweka.elastic_distance_measures.DTW_DistanceBasic;
 import java.util.HashMap;
-import evaluation.ClassifierResults;
+import evaluation.storage.ClassifierResults;
 import utilities.ClassifierTools;
-import evaluation.CrossValidator;
+import evaluation.evaluators.CrossValidationEvaluator;
 import utilities.TrainAccuracyEstimate;
 import vector_classifiers.SaveEachParameter;
 import weka.classifiers.AbstractClassifier;
