+++ src/main/java/vector_classifiers/TunedRandomForest.java
@@ -37,12 +37,12 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import timeseriesweka.classifiers.ParameterSplittable;
 import utilities.ClassifierTools;
-import evaluation.CrossValidator;
+import evaluation.evaluators.CrossValidationEvaluator;
 import utilities.InstanceTools;
 import timeseriesweka.classifiers.SaveParameterInfo;
 import utilities.TrainAccuracyEstimate;
 import weka.classifiers.meta.Bagging;
-import evaluation.ClassifierResults;
+import evaluation.storage.ClassifierResults;
 import static utilities.GenericTools.indexOfMax;
 import weka.classifiers.trees.RandomForest;
 import weka.classifiers.trees.RandomTree;
@@ -261,7 +261,7 @@ public class TunedRandomForest extends RandomForest implements SaveParameterInfo
         double minErr=1;
         this.setSeed(rng.nextInt());
         Instances trainCopy=new Instances(train);
-        CrossValidator cv = new CrossValidator();
+        CrossValidationEvaluator cv = new CrossValidationEvaluator();
         cv.setSeed(seed);
         cv.setNumFolds(folds);
         cv.buildFolds(trainCopy);
@@ -475,7 +475,7 @@ public class TunedRandomForest extends RandomForest implements SaveParameterInfo
                 t.setNumFeatures(this.getNumFeatures());
                 t.setNumTrees(this.getNumTrees());
                 t.setSeed(seed);
-                CrossValidator cv = new CrossValidator();
+                CrossValidationEvaluator cv = new CrossValidationEvaluator();
                 cv.setSeed(seed); 
                 cv.setNumFolds(folds);
                 cv.buildFolds(data);
