+++ src/main/java/vector_classifiers/weightedvoters/CAWPE_TunedAlpha.java
@@ -8,7 +8,7 @@ import timeseriesweka.classifiers.ensembles.voting.BestIndividualTrain;
 import timeseriesweka.classifiers.ensembles.voting.MajorityConfidence;
 import timeseriesweka.classifiers.ensembles.weightings.EqualWeighting;
 import timeseriesweka.classifiers.ensembles.weightings.TrainAcc;
-import evaluation.ClassifierResults;
+import evaluation.storage.ClassifierResults;
 import utilities.StatisticalUtilities;
 import vector_classifiers.CAWPE;
 import weka.core.Instances;
